third_party/blink/web_tests/platform/generic/fast/dom/Window/mozilla-focus-blur-expected.txt

Check that window.blur() does nothing, and window.focus() only works if it is invoked from the window that opened the former. If the test passes, you should see a series of PASS messages with the last being 'All tests finished'.

This test is adopted from mozilla's tests.

PASS: The focus should not have been changed!
PASS: The focus should not have been changed!
PASS: The focus should not have been changed with URL=resources/mozilla-focus-blur-popup-opener-focus.html
PASS: The focus should not have been changed with URL=resources/mozilla-focus-blur-popup-blur.html
PASS: The last opened window should be able to get focus
PASS: All tests finished
